refactor: removable textfield 삭제
